feat: Create separate schemas for US and UK
